pyproject.toml
[project]
name = "server-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.4.0",
    "pathlib>=1.0.1",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "streamlit>=1.50.0",
    "wordcloud>=1.9.4",
]
